+++ software/apps/baremetal/softmax_f16/main.c
@@ -2,8 +2,6 @@
 // Licensed under the Apache License, Version 2.0, see LICENSE for details.
 // SPDX-License-Identifier: Apache-2.0
 
-// Author: Samuel Riedel, ETH Zurich
-
 #include <stdint.h>
 #include <string.h>
 
@@ -13,24 +11,14 @@
 #include "runtime.h"
 #include "synchronization.h"
 
-#include "data_matmul_f16.h"
+#include "data_softmax_f16.h"
 
 #include "baremetal/mempool_checks.h"
-#include "baremetal/mempool_matmul_f16.h"
-
-/*
-======================
-Parameters and defines
-
-SINGLE: When defined runs single-core matmul.
-PARALLEL: When defined runs parallel matmul.
-*/
+#include "baremetal/mempool_softmax_f16.h"
 
 __fp16 matrix_a[matrix_M * matrix_N]
     __attribute__((aligned(sizeof(int32_t)), section(".l1_prio")));
-__fp16 matrix_b[matrix_N * matrix_P]
-    __attribute__((aligned(sizeof(int32_t)), section(".l1_prio")));
-__fp16 matrix_c[matrix_M * matrix_P]
+__fp16 matrix_b[matrix_M * matrix_N]
     __attribute__((aligned(sizeof(int32_t)), section(".l1_prio")));
 
 int main() {
@@ -39,36 +27,23 @@ int main() {
   // Initialize barrier and synchronize
   mempool_barrier_init(core_id);
 
-  // Initialize Matrices 1
+  // Initialize matrices
   if (core_id == 0) {
     dma_memcpy_blocking(matrix_a, l2_A,
                         (matrix_M * matrix_N) * sizeof(int16_t));
     dma_memcpy_blocking(matrix_b, l2_B,
-                        (matrix_N * matrix_P) * sizeof(int16_t));
+                        (matrix_M * matrix_N) * sizeof(int16_t));
   }
   mempool_barrier(num_cores);
 
-#if defined(SINGLE)
-  if (core_id == 0) {
-    // Execute function to test.
-    mempool_start_benchmark();
-    matmul_2x2_single_f16(matrix_a, matrix_b, matrix_c, matrix_M, matrix_N,
-                          matrix_P);
-    mempool_stop_benchmark();
-  }
-  mempool_barrier(num_cores);
-#endif
-
-#if defined(PARALLEL)
-  // Execute function to test.
+  // Matrix Normalization
   mempool_start_benchmark();
-  matmul_4x2_parallel_f16vec(matrix_a, matrix_b, matrix_c, matrix_M, matrix_N,
-                             matrix_P, core_id, num_cores);
+  softmax_parallel_2x4_f16vec(matrix_a, matrix_b, matrix_M, matrix_N, core_id,
+                              num_cores);
   mempool_barrier(num_cores);
   mempool_stop_benchmark();
-#endif
 
-  mempool_check_f16(matrix_c, l2_C, matrix_M * matrix_P, 0.5f, 0);
+  // mempool_check_f16(matrix_b, l2_B, matrix_M * matrix_N, 0.1f, 1);
   mempool_barrier(num_cores);
   return 0;
 }